make TestSwitchStepperIcon own component
